Toasts
Add LabelHandler
Remove GetWrappedLabelText()
#73
